requirements.txt
undetected-chromedriver
selenium
python-telegram-bot==13.15
requests
Flask
